Create seed data for testing expenses reducer
