+++ src/tests/reducers/expenses.test.js
@@ -1,4 +1,5 @@
 import expensesReducer from '../../reducers/expenses';
+import expenses from '../fixtures/expenses';
 
 test('should set default state', () => {
 	const state = expensesReducer(undefined, {type: '@@INIT'});
