+++ src/tests/selectors/expenses.test.js
@@ -1,30 +1,7 @@
 import moment from 'moment';
 import selectExpenses from '../../selectors/expenses';
+import expenses from '../fixtures/expenses';
 
-// expense data for testing filters
-const expenses = [
-	{
-		id: '1',
-		description: 'Gummy',
-		note: '',
-		amount: 195,
-		createdAt: 0
-	},
-	{
-		id: '2',
-		description: 'Rent',
-		note: '',
-		amount: 109500,
-		createdAt: moment(0).subtract(4, 'days').valueOf()
-	},
-	{
-		id: '3',
-		description: 'Credit Card',
-		note: '',
-		amount: 4500,
-		createdAt: moment(0).add(4, 'days').valueOf()
-	}
-];
 test('Test should filter by text value', () => {
 	const filters = {
 		text: 'e',
